made mobile header
